Menu background
I'll use this commit as a mini blog post because it's 6:20 AM and
I need to express myself. I was a bit depressed while writing the UI
code, especially with the plain white background, and I came to a theory
that justifies the order in which I do things sometimes.

You see, doing something ugly eats you out. So yes, in theory, you
should get the gameplay, main mechanisms, stats, interactions,
strategies, behaviors, etc. working first - but when working on stuff
like the UI - having a placeholder UI just kills me.

Is it because I care too much about what I do, even when it's a project
for fun, of no actual importance? Probably. If you love what you do, you
can't stand seeing it in a state of ugliness. So even if you should (and
should is a bad word) be doing tons of other stuff, sometimes you just
stop and make it prettier, even if it delays other stuff.

That also applies to non-game-programming stuff in life, btw. Thanks for
reading.